src/tests/graphql-aspnet-tests/Parsing2/NodeBuilders/Inputs/EnumValueNodeBuilderTests.cs
// *************************************************************
// project:  graphql-aspnet
// --
// repo: https://github.com/graphql-aspnet
// docs: https://graphql-aspnet.github.io
// --
// License:  MIT
// *************************************************************

namespace GraphQL.AspNet.Tests.Parsing2.NodeBuilders.Inputs
{
    using System;
    using GraphQL.AspNet.Parsing2.Lexing;
    using GraphQL.AspNet.Parsing2.Exceptions;
    using GraphQL.AspNet.Parsing2.Lexing.Source;
    using GraphQL.AspNet.Parsing2.Lexing.Tokens;
    using GraphQL.AspNet.Parsing2;
    using GraphQL.AspNet.Parsing2.NodeBuilders.Inputs;
    using GraphQL.AspNet.Tests.CommonHelpers;
    using NUnit.Framework;
    using GraphQL.AspNet.Tests.Parsing2.Helpers;

    [TestFixture]
    public class EnumValueNodeBuilderTests
    {
        [Test]
        public void EnumValueNodeMaker_PointingAtAName_ReturnsEnum()
        {
            var text = "JEDI, name = 343";
            var stream = Lexer.Tokenize(new SourceText(text.AsSpan()));
            stream.Prime();

            var tree = SynTree.FromDocumentRoot();
            var docNode = tree.RootNode;

            EnumValueNodeBuilder.Instance.BuildNode(ref tree, ref docNode, ref stream);

            HelperAsserts.AssertChildNodeChain(
                stream.Source,
                tree,
                docNode,
                new SynNodeTestCase(
                    SynNodeType.EnumValue,
                    "JEDI"));

            // ensure stream is pointing at "name"
            Assert.AreEqual(TokenType.Name, stream.TokenType);
            SynTreeOperations.Release(ref tree);
        }

        [Test]
        public void EnumValueNodeMaker_WillInterpreteABoolAsEnum_IfInvoked()
        {
            var text = "true, name = 343";
            var stream = Lexer.Tokenize(new SourceText(text.AsSpan()));
            stream.Prime();

            var tree = SynTree.FromDocumentRoot();
            var docNode = tree.RootNode;

            EnumValueNodeBuilder.Instance.BuildNode(ref tree, ref docNode, ref stream);

            HelperAsserts.AssertChildNodeChain(
                stream.Source,
                tree,
                docNode,
                new SynNodeTestCase(
                    SynNodeType.EnumValue,
                    "true"));

            // ensure stream is pointing at "name"
            Assert.AreEqual(TokenType.Name, stream.TokenType);
            SynTreeOperations.Release(ref tree);
        }

        [Test]
        public void EnumValueNodeMaker_NotPointingAtAName_ResultsInException()
        {
            var text = "123, arg2: \"stringValue\")";
            var stream = Lexer.Tokenize(new SourceText(text.AsSpan()));
            stream.Prime();

            var tree = SynTree.FromDocumentRoot();
            var docNode = tree.RootNode;
            try
            {
                EnumValueNodeBuilder.Instance.BuildNode(ref tree, ref docNode, ref stream);
            }
            catch (GraphQLSyntaxException)
            {
                return;
            }
            finally
            {
                SynTreeOperations.Release(ref tree);
            }

            Assert.Fail("Expection syntax exception");
        }
    }
}